multistrap: Ignore native arch
Multistrap supports two modes:
 - Native: target and host architectures are the same.
 - Cross: target architecture differs from host.

In native mode, multistrap automatically runs pos-install scripts for
all the packages installed to rootfs. This could be a problem for Isar
build flow, becuase configscript.sh assumes, that rootfs is not
configured. Also this leads to different Isar flows for native and cross
architectures what should not happen.

Signed-off-by: Alexander Smirnov <asmirnov@ilbers.de>